Clarify what parentId is meant to do
The case for parentId was ambiguous. It has now been clarified that
parentId it meant to communicate the reason why a resource was created.
In contrast to currentId(), which communicates when a resource is being
created.

Added two examples to further clarify the use case for parentId.

Add note in destroy() that GC'd handles need to manually call destroy()
when its operation is complete.

Grammatical fixes and clarifications.