Rename parentId to triggerId
Intuitively the "parentId" of a given execution scope would be the
"currentId" of the execution scope that created the handle that's
currently executing.

Using the formal definition for trigger "cause (an event or situation)
to happen or exist", the "triggerId" is the id of the handle responsible
for init() being called. For example:

    const server = net.createServer(conn => {
      // currentId() returns the id of the handle that's executing
      // the callback. In this case it's the id of "server".
      async_hooks.currentId();

      // Though the handle that caused (or triggered) this callback to
      // be called was that of the new connection. Thus the return value
      // of triggerId() is the id of "conn".
      async_hooks.triggerId();
    }).listen();

Add async_hooks.triggerId() API to return the trigger id responsible for
the current execution scope.